Update style.css
im not front developer but there is a bug.

fix: Add scrollbar to tags container in the Random Choice Picker

Previously, when a large amount of data was entered into the textarea, causing numerous tags to be generated, the tags container would expand without providing scrollbars. This commit adds a maximum height to the tags container and enables vertical scrolling when the content exceeds this height. Now, users can scroll through the tags when there are many of them, improving the usability of the application.